Separate server and client code to fix build errors.
