+++ src/app/(main)/admin/users/[id]/edit/actions.ts
@@ -4,7 +4,8 @@ import { z } from 'zod'
 import bcrypt from 'bcrypt'
 import prisma from '@/lib/prisma'
 import { revalidatePath, revalidateTag } from 'next/cache'
-import { uploadFile, deleteFile, getPublicUrl } from '@/lib/storage'
+import { getPublicUrl } from '@/lib/storage'
+import { uploadFile, deleteFile } from '@/lib/actions/storage'
 import { getCodeTable } from '@/lib/codes'
 import { auth } from '@/auth'
 import { Gender, DatePrecision, User } from '@/generated/prisma/client'
