+++ src/app/(main)/admin/users/create/actions.ts
@@ -4,7 +4,7 @@ import { z } from 'zod'
 import prisma from '@/lib/prisma'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
-import { deleteFile, uploadFile } from '@/lib/storage'
+import { deleteFile, uploadFile } from '@/lib/actions/storage'
 import { getCodeTable } from '@/lib/codes'
 import { auth } from '@/auth'
 import bcrypt from 'bcrypt'
