+++ src/app/(main)/me/users/actions.ts
@@ -11,7 +11,8 @@ import {
 import { auth } from '@/auth'
 import prisma from '@/lib/prisma'
 import bcrypt from 'bcrypt'
-import { uploadFile, getPublicUrl, deleteFile } from '@/lib/storage'
+import { getPublicUrl } from '@/lib/storage'
+import { uploadFile, deleteFile } from '@/lib/actions/storage'
 import { revalidatePath } from 'next/cache'
 import crypto from 'crypto'
 import { parseDateAndDeterminePrecision } from '@/lib/utils'
