+++ src/app/g/[slug]/admin/actions.ts
@@ -5,7 +5,7 @@ import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
 import { z } from 'zod'
 
-import { uploadFile, deleteFile } from '@/lib/storage'
+import { uploadFile, deleteFile } from '@/lib/actions/storage'
 import { auth } from '@/auth'
 import { getCodeTable } from '@/lib/codes'
 import { isAdmin } from '@/lib/auth-utils'
